+++ packages/cli/CHANGELOG.md
@@ -15,12 +15,14 @@
 - Added a `shell` task option to `moon.yml` that will wrap system tasks in a shell.
 - Added a `tags` setting to `moon.yml` for project categorization.
 - Added a `--tags` option to the `moon query projects` command.
+- Added a `telemetry` setting to `.moon/workspace.yml`.
 - Added 5 new token variables: `$projectAlias`, `$date`, `$time`, `$datetime`, and `$timestamp`.
 - Updated task `env` and `.env` files to support variable substitution using `${VAR_NAME}` syntax.
 - Updated system tasks to now execute within a shell.
 
 #### ⚙️ Internal
 
+- Updated the new version check to only run on the `check`, `ci`, `run`, and `sync` commands.
 - Will now detect 16 additional CI environments: Agola, AppCenter, Appcircle, Azure, Bamboo,
   Bitrise, Buddy, Cirrus, Codemagic, Heroku, Jenkins, Jenkins X, Netlify, TeamCity, Vela,
   Woodpecker.